# app/api/resend/webhook/route.ts
import { NextResponse, type NextRequest } from "next/server"

import { parseResendDeliveryEvent } from "@/lib/loyalty-invites/webhook-core"
import { verifyStandardWebhook } from "@/lib/notifications/standard-webhook"
import { logger } from "@/lib/observability/logger"
import { createSupabaseServiceRoleClient } from "@/lib/supabase/server"

/**
 * Signed, replay-safe Resend delivery webhook for bulk loyalty invitations.
 *
 * Resend signs with the Standard Webhooks (Svix) scheme, so verification reuses
 * verifyStandardWebhook (HMAC over `${id}.${timestamp}.${body}`, 5-minute replay
 * window). Delivered/bounced/complained/failed events are applied idempotently
 * to the recipient by provider message id; bounces and complaints also suppress
 * the address globally. It fails closed when RESEND_WEBHOOK_SECRET is unset.
 */

export const runtime = "nodejs"
export const dynamic = "force-dynamic"

const NO_STORE = { "cache-control": "no-store, max-age=0" }

export async function POST(request: NextRequest) {
  const secret = process.env.RESEND_WEBHOOK_SECRET?.trim()
  if (!secret) {
    logger.warn("loyalty_invite_webhook_unconfigured", {})
    return NextResponse.json(
      { error: "unconfigured" },
      { status: 503, headers: NO_STORE }
    )
  }

  const body = await request.text()
  const verified = verifyStandardWebhook({
    secret,
    id: request.headers.get("webhook-id") ?? "",
    timestamp: request.headers.get("webhook-timestamp") ?? "",
    body,
    signatureHeader: request.headers.get("webhook-signature") ?? "",
  })
  if (!verified) {
    return NextResponse.json(
      { error: "invalid_signature" },
      { status: 401, headers: NO_STORE }
    )
  }

  let payload: unknown
  try {
    payload = JSON.parse(body)
  } catch {
    return NextResponse.json(
      { error: "invalid_payload" },
      { status: 400, headers: NO_STORE }
    )
  }

  const parsed = parseResendDeliveryEvent(payload)
  if (!parsed) {
    // Acknowledge unhandled event types (sent/opened/clicked/delivery_delayed).
    return NextResponse.json({ ok: true, ignored: true }, { headers: NO_STORE })
  }

  const supabase = createSupabaseServiceRoleClient()
  const { error } = await supabase.rpc("apply_loyalty_invite_delivery_event", {
    p_provider_message_id: parsed.providerMessageId,
    p_event: parsed.event,
  })
  if (error) {
    logger.warn("loyalty_invite_webhook_apply_failed", {
      reason: "database_rejected",
    })
    return NextResponse.json(
      { error: "apply_failed" },
      { status: 500, headers: NO_STORE }
    )
  }

  return NextResponse.json({ ok: true }, { headers: NO_STORE })
}